requirements.txt
discord-py>=2.5.2
flask>=3.1.1
psycopg2-binary>=2.9.9
